fix distance

+++ mogate.Shared/Actions/Gameplay/TriggerEntity.cs
@@ -17,7 +17,7 @@ namespace mogate
 		public bool Execute(GameTime gameTime)
 		{
 			if (m_to.Has<Triggerable> ()) {
-				int dist = (int)Utils.Dist (m_from.Get<Position> ().MapPos, m_to.Get<Position> ().MapPos);
+				int dist = Utils.Dist (m_from.Get<Position> ().MapPos, m_to.Get<Position> ().MapPos);
 				if (dist <= m_to.Get<Triggerable> ().Distance)
 					m_to.Get<Triggerable> ().Trigger (m_from, m_to);
 			}
