+++ mogate.Shared/Core/Utils.cs
@@ -15,9 +15,9 @@ namespace mogate
 			return Enum.GetValues(typeof (T)).Cast<T>().OrderBy(x => Rand.Next()).FirstOrDefault();
 		}
 
-		public static double Dist (Point from, Point to)
+		public static int Dist (Point from, Point to)
 		{
-			return Math.Sqrt(Math.Pow(from.X  - to.X, 2) + Math.Pow(from.Y - to.Y, 2));
+			return (int)Math.Sqrt(Math.Pow(from.X  - to.X, 2) + Math.Pow(from.Y - to.Y, 2));
 		}
 
 		public static bool DropChance(int chance)
@@ -30,16 +30,16 @@ namespace mogate
 			return Rand.Next (dice);
 		}
 
-		public static double DirectionDist (Point from, Point to, Direction fromDir)
+		public static int DirectionDist (Point from, Point to, Direction fromDir)
 		{
 			if (fromDir == Direction.Up && (from.X != to.X || from.Y < to.Y))
-				return double.MaxValue;
+				return int.MaxValue;
 			if (fromDir == Direction.Down && (from.X != to.X || from.Y > to.Y))
-				return double.MaxValue;
+				return int.MaxValue;
 			if (fromDir == Direction.Left && (from.X < to.X || from.Y != to.Y))
-				return double.MaxValue;
+				return int.MaxValue;
 			if (fromDir == Direction.Right && (from.X > to.X || from.Y != to.Y))
-				return double.MaxValue;
+				return int.MaxValue;
 			return Dist (from, to);
 		}
